Add a modified simplify function for building Hamiltonians (#2103)
* add modified simplify function
